backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206
opencv-python==4.8.1.78
pillow==10.0.1
PyMuPDF==1.23.8
numpy==1.25.2
pandas==2.1.3